Changes knight armour slowdown
Changed the knight armours to have a slowdown of 0.5 instead of 0, and the robin version to have a slowdown of 0 rather than -1. On this server these are mostly used just as fluff items for the fantasy areas, and people have been rushing to the robin armour in the fantasy redgate due to its movement speed boost. Rather than removing the armour from the map, making it not a straight up movement buff and letting people still find it is better, as it's not used elsewhere.