Fixes cave-ins in fantasy dungeon (#17667)
* Changes knight armour slowdown

Changed the knight armours to have a slowdown of 0.5 instead of 0, and the robin version to have a slowdown of 0 rather than -1. On this server these are mostly used just as fluff items for the fantasy areas, and people have been rushing to the robin armour in the fantasy redgate due to its movement speed boost. Rather than removing the armour from the map, making it not a straight up movement buff and letting people still find it is better, as it's not used elsewhere.

* Fantasy redgate changes

Added new variants of knight armours that have lower stats and small slowdown, but can be found around the fantasy redgate instead of the buffed PoI rewards.

Replaced the murder fire smoke from the burning building with harmless smoke.

* Fixes cave-ins in fantasy dungeon

Fixed cave-ins in fantasy dungeon around the dragon