lets you make drying racks using steel
The lazy solution.

Look, it's a smart fridge, I don't know why. I think it's silly that it's a smart fridge. whenever you put something in there it is instantly dried. if I wasn't making this PR right before I go to sleep, I would make it so you can dry plants on the tanning rack. or making tanning rack structure called a drying rack with some changed around sprites. But that requires being smart, and I am not that when midnight looms. So to motivate myself I am going to just cold PR this with a DNM so I am reminded to fix this.

I mean for fucks sake the thing isn't even deconstructable.