Torso definition markings
Added a bunch of new torso definition markings: Bust/Bustier (solid, translucent and outline), nipples (normal position, closer together and one pixel lower) and abs.

This is because some species already have these features, but people using markings for chest fur or starting as a species without them (such as unathi) find it frustrating that they lose these features.
The solid version is better for those that want to hide the definition of the body beneath them for smoother chests, translucent is better for people layering it with other markings, and outline only for a weaker definition.