Made backup batteries a bit more useful
Changed the level that you can use backup batteries as a synth to 150 nutrition, and it also gives you 150 nutrition back. This means you can use it as soon as you reach slowdown levels of hunger and it fills you up a decent amount. Previously you had to wait until you went down to 100 nutrition and it would only fill up to 200 max, which means you'd be back in red soon after.