Containment field exploit fix
Fixed an exploit that allowed dense objects to be pushed in front of the containment field when it was already active. It already had a check on startup to prevent starting with a dense object in the way, but did not check afterwards. It now instantly shuts off when a dense object crosses over it.