Alternate base sprites for zorren (#18093)
* [WIP] Alternate base sprites for zorren

Currently does not work, may need some help to figure out why.

The idea is to change the default base sprite of the zorren to a much lighter one, as many people complain about the dark greyscaling on the original. However, to preserve people's characters who incorporate the current sprites, I wanted to make this an option that could be selected through the custom species icon menu.

If I disable the "selects bodytype" var for the species, then the new light version appears as you would expect and everything works fine.
However, if I enable the new zorren define for selects bodytype, it gives you the two options in the menu as you would expect, but the appearance for both types is the dark version ingame. Basically, this does not seem to work and I can't get my head around why.

* Actually works now!