Gold Star Printer (#18129)
* Gold Star Printer

Added a new Gold Star Printer to HoP offices. This printer prints gold stars with a short title of "Gold Star for X" and a custom description for more info. These gold stars can be used as accessories on underclothing. They can also be stuck to ANY object that doesn't have a custom attack interaction (so, it won't stick to tables, or most machines that you can put things in), and when stuck to an object it simply adds a little suffix to it's description stating it has a gold star attached. Yes, you can attach it to absurd objects but that's funny.

Any printed stickers are admin logged so we can track abusive stickers.

* Removes sanitize