Posing buffs (#18433)
* Posing buffs

Changed posing to now post the posing message in chat when initiated.

Added a posing indicator with a colour determined by runechat colour when a pose is active.

Added an option to have a pose vanish immediately upon movement.

* Light animation

* Allows more options

UGH, checkboxes are so good for this but it's a pain that they don't work with the framework turned off.

* Pretties up the tgui inputs

* Update emote.dm

---------

Co-authored-by: Kashargul <144968721+Kashargul@users.noreply.github.com>