Revamps flash code (#18743)
* Revamps flash code

Makes flash code less crungo by making it a proc that returns true/false instead of a giant if() chain.

Makes flashes be able to be repaired by default. It seems like this feature was disabled due to a 'compile error' (????) but it doesn't happen anymore, so we can let it back.

* Update flash.dm

* no spawn