Makes sizeoxadone more affordable
Changes the cost of sizeoxadone from (1 clonexadone, 3 tramadol and 1 phoron) to (1 sodium and 1 tramadol). It still uses phoron as a catalyst but that is not used up in the reaction.

The idea behind this is to eliminate the extremely high phoron cost of making size chems. There's already so many totally free ways to change your own or someone else's size that it makes no sense that you need to deplete half of the chemistry labs phoron stock to flavour that as a chemical effect.