Laserpointer changes
Cleans up laser pointer code a bit
Makes it so having TOO STRONG of a laser won't break it
Makes it so laser pointering someone without eyes doesn't return early and break the entire chain
Makes it so that being blind (or otherwise unable to see) makes it so you can't uber blind someone
Makes it so that using a laser pointer (no matter the strength) on someone with FLASH_PROTECTION_MAJOR (dispersed eyes, welding goggles, welding mask, etc) does not affect them

Fixes a bug where lasers that were too strong would fail to EMP a camera due to division by 0 error.

Gets rid of the double-RNG with laser pointers when it comes to hitting a camera.